refactor: Move symbol use and redeclaration checks to the symbol table.
Allow definition of symbols to insert multiple values during parse, and worry
about conflicts later on.

This will have flow on effects for handling `subst()` parameters later on.

We check in the scope `Check` method now for redeclaration of variables when
there are more than one definition, and we continue to check for use of any
variable when defined.

Add `Equal` methods to `Position` and `Error` so that they can be compared in
test.